datapath: move interface definitions to /types
In preparation of moving the datapath module into datapath/cells.go, we
need to break import cycles by moving the interface definitions into
their own package. Reuse the existing datapath/types package for this,
and adjust code and tests everywhere accordingly.
