+++ pkg/datapath/loader/hash_test.go
@@ -6,8 +6,8 @@ package loader
 import (
 	. "gopkg.in/check.v1"
 
-	"github.com/cilium/cilium/pkg/datapath"
 	"github.com/cilium/cilium/pkg/datapath/linux/config"
+	datapath "github.com/cilium/cilium/pkg/datapath/types"
 	"github.com/cilium/cilium/pkg/testutils"
 )
