+++ pkg/datapath/types/wireguard.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package datapath
+package types
 
 import (
 	"net"
